+++ lib/core/analytics/analytics_provider.dart
@@ -1,0 +1,7 @@
+import 'package:agricola_dashboard/core/analytics/analytics_service.dart';
+import 'package:firebase_analytics/firebase_analytics.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+final analyticsServiceProvider = Provider<AnalyticsService>(
+  (ref) => AnalyticsService(FirebaseAnalytics.instance),
+);